requirements.txt
<<<<<<< Updated upstream
Django==1.9
Pillow==2.3.0
SecretStorage==2.0.0
argparse==1.2.1
colorama==0.2.5
django-ckeditor==5.3.0
django-js-asset==0.1.1
html5lib==0.999
httplib2==0.8
requests==2.2.1
simplejson==3.3.1
six==1.5.2
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wheel==0.24.0
=======
Django==1.9
Pillow==2.3.0
SecretStorage==2.0.0
argparse==1.2.1
colorama==0.2.5
django-ckeditor==5.3.0
django-js-asset==0.1.1
html5lib==0.999
httplib2==0.8
requests==2.2.1
simplejson==3.3.1
six==1.5.2
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wheel==0.24.0
>>>>>>> Stashed changes
gunicorn